// frontend/src/types/generated/ServerProductSurface.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HttpRouteSurface } from "./HttpRouteSurface.js";
import type { RoomAction } from "./RoomAction.js";
import type { RoomStream } from "./RoomStream.js";

export type ServerProductSurface = { revision: number, digest: string, http_routes: Array<HttpRouteSurface>, websocket_streams: Array<RoomStream>, websocket_actions: Array<RoomAction>, };